pandityatra/backend/requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
dj-database-url
djangorestframework-simplejwt
python-dotenv

# Django Channels & WebSocket support
channels[daphne]
channels-redis

# Redis for channel layers
redis

# Timezone support
pytz

# Payment gateways
stripe
requests  # For Khalti/eSewa API calls

# Translation
django-modeltranslation

# OpenAI API for AI guide
openai

# Video call integration

requests

# For handling CORS
django-cors-headers 

# For kundali generation
pyswisseph  

# For handling file uploads
django-storages
boto3  # If using AWS S3 for storage

dj-database-url

pillow